refactor(operator): extract pod_builder helpers; squad pods get full env (secrets, MCP, OAuth)
